Исправления.

+++ BookStore/src/main/java/ru/otus/spring/bookstore/restcontroller/AutorController.java
@@ -1,18 +1,27 @@
 package ru.otus.spring.bookstore.restcontroller;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+import org.springframework.web.server.handler.ExceptionHandlingWebHandler;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import ru.otus.spring.bookstore.domain.Book;
+import ru.otus.spring.bookstore.repository.BookRepository;
 import ru.otus.spring.bookstore.restcontroller.dto.AutorDto;
 import ru.otus.spring.bookstore.domain.Autor;
 import ru.otus.spring.bookstore.repository.AutorRepository;
 
+import java.net.http.HttpResponse;
+
 @RestController
 public class AutorController {
     private final AutorRepository autorRepository;
+    private final BookRepository bookRepository;
 
-    public  AutorController(AutorRepository autorRepository){
+    public AutorController(AutorRepository autorRepository, BookRepository bookRepository) {
         this.autorRepository = autorRepository;
+        this.bookRepository = bookRepository;
     }
 
     @GetMapping(value = "/api/autor", params = "name")
@@ -44,10 +53,18 @@ public class AutorController {
 
     }
 
+
+
     @DeleteMapping(path="/api/autor")
     public Mono<Void> delete(@RequestBody AutorDto autorDto){
         Autor autor = AutorDto.toAutor(autorDto);
-        return autorRepository.delete(autor);
+        return bookRepository.existsByAutor(autor).flatMap(b -> {
+                                                                    if(!b) {
+                                                                        return autorRepository.delete(autor);
+                                                                    }else {
+                                                                       throw new ResponseStatusException(HttpStatus.FOUND, "У автора имеются книги");
+                                                                    }
+                                                                } );
     }
 
 }
